Agregacion de estilos, modal de confirmacion y acomodacion de la pagina

+++ backend/Models/Historial.cs
@@ -5,9 +5,9 @@
         public int Id { get; set; }
         public int UsuarioId { get; set; }
         public int MonedaId { get; set; }
-        public double Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public DateTime Fecha { get; set; }
-        public double Cotizacion { get; set; }
+        public decimal Cotizacion { get; set; }
         public Usuario? Usuario { get; set; }
         public Moneda? Moneda { get; set; }
     }
